requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
asyncio-dgram==2.2.0
attrs==25.3.0
discord.py==2.5.2
dnspython==2.7.0
frozenlist==1.7.0
idna==3.10
mcstatus==12.0.2
multidict==6.5.0
propcache==0.3.2
python-dotenv==1.1.0
typing_extensions==4.14.0
yarl==1.20.1
